Address Sonar issues: - Extract provided builders of AbstractItem to a trait - Rename local variable in DefaultMapper
